+++ CHANGELOG.md
@@ -6,7 +6,8 @@
   * `MicroUtils`: `0.19.2`
   * `PlaguBot`: `6.0.1`
   * `TelegramBotAPILibraries`: `0.12.1`
->>>>>>> update dependnecies
+* `Captcha`:
+  * Add support of positive users (task [#31](https://github.com/InsanusMokrassar/PlaguBotPlugins/issues/31))
 
 ## 0.11.6